Added fullscreen and mouse corrected

+++ include/objects/Camera.cpp
@@ -163,7 +163,7 @@ void Camera::rotateCenter(float ang_xy, float ang_z, float length) {
     //	view = lookAt(vec3(cam.getPos()), vec3(cam.getPosFocus()), vec3(0.0, 0.0, 1.0));
 }
 
-void Camera::rotateFoco(float ang_xy, float ang_z, double timeDif) {
+void Camera::rotateFoco(double ang_xy, double ang_z, double timeDif) {
     eye = glm::vec3(
             glm::vec4(eye, 1.0f) *
             rotate(glm::mat4(1.0f), (float) (ang_xy * camSensibility * timeDif), glm::vec3(0, 0, 1)));
