+++ include/objects/Camera.hpp
@@ -58,7 +58,7 @@ struct Camera {
     void move(glm::vec3 v);
     void moveFPS(float f_b, float r_l, float u_d, double timeDif);
     void rotateCenter(float ang_xy, float ang_z, float length);
-    void rotateFoco(float ang_xy, float ang_z, double timeDif);
+    void rotateFoco(double ang_xy, double ang_z, double timeDif);
     void pan(float x, float y);
 
     void linkCamAttrib(GLuint shader);
